Move `IconLayers` to its own package
