+++ src/main/kotlin/character/Guild.kt
@@ -1,6 +1,6 @@
 package cloud.drakon.ktlodestone.character
 
-import cloud.drakon.ktlodestone.IconLayers
+import cloud.drakon.ktlodestone.iconlayers.IconLayers
 
 /** A [Character]'s [FreeCompany] or [PvpTeam]. */
 data class Guild(val name: String, val id: String, val iconLayers: IconLayers)
