+++ src/main/kotlin/iconlayers/IconLayers.kt
@@ -1,3 +1,3 @@
-package cloud.drakon.ktlodestone
+package cloud.drakon.ktlodestone.iconlayers
 
 data class IconLayers(val bottom: String, val middle: String, val top: String)
